Updating of 10 Models & Uploading of relevant textures
An update on the 10 models I created last Trimester after new experience with unity and Dennis Yu brought issues with the models to my attention on the 12th. The issues were overlapping and disappearing geometry and the textures were not applied to the FBX files. I have updated the models geometry and textures, along with adding the textures to the textures folder and making the path between them relative so that the textures should stick.