Update to AngleControl
Update to AngleControl with code that causes the calculation for the angle and the publishing of that data to occur when the player collides with something and the player remains in contact with the terrain, so long as the player has a non-kinematic rigidbody. Should fix the issues Ethan is having with testing though I don't know why lack of collision triggers would affect the code since it no longer uses collisions in the calculations.