Update to AngleControl, removing collision triggers for script
Update to AngleControl, removing collision triggers for script and adding a MQTT object to the scene to attach the MQTT script to. Also added a piece of code that causes the AngleControl script to write the payload to the debug log to confirm the code is functioning. Could not set up the MQTT for testing as I was missing the device id and certificate.