Update PlayerController.cs
To those who come after us.

This PR is where we need to go forward for the scene interaction for the bike, but due to the changes between how this works with an engine instead of direct control of velocity the MQTT system might not work with this.

MQTT may need to use the power value rather than the speed to apply to the bike. As this is so close to the end of trimester and wanting to have a MVP build, this will be left as a PR for those picking up and taking over next trimester.